refactor(attack-paths): address PR review feedback
- Extract GRAPH_ALERT_BORDER_COLOR constant to graph-colors.ts
- Extract getPathEdges function to graph-utils.ts for better reusability
- Replace hardcoded #ef4444 color values with the extracted constant
- Add comment explaining D3 type cast for clarity
