+++ ui/app/(prowler)/attack-paths/(workflow)/query-builder/_lib/index.ts
@@ -4,9 +4,11 @@ export {
   exportGraphAsSVG,
 } from "./export";
 export { formatNodeLabel, formatNodeLabels } from "./format";
+export { getPathEdges } from "./graph-utils";
 export {
   getNodeBorderColor,
   getNodeColor,
+  GRAPH_ALERT_BORDER_COLOR,
   GRAPH_EDGE_COLOR,
   GRAPH_EDGE_HIGHLIGHT_COLOR,
   GRAPH_NODE_BORDER_COLORS,
